Changed the way to detect Changes to Files.
Instead of Checking if steam is running, and changing it from that, I
have it set now to set the custom wallpaper, then check if it is ever
changed, and change it back. Currently has a lot of problems, one being
that it constantly detects that it is changing the file, and changes
itself. Additionally, I still can not get it to work when steam launches
directly into Big Picture, a situation that I believe may be a more
often use case than launching the desktop app, then going into Big
Picture Mode.